[Core] Added TextureGroup; Fixes #35; [Core] Removed unused files; [Core] Fixed some drawing issues on SFMLRenderer; [Core] Added platform ID info on GameInterface; [Core] Added GameInterface::GraphicsEnabled flag; [Core] GameInterface::IsMobile is now modifiable;

+++ Development/Core/StringUtils.cpp
@@ -48,7 +48,7 @@ namespace FlamingTorch
 		if(EndIndex == -1)
 			return std::string();
 
-		while(StartIndex < str.length())
+		while(StartIndex < (int32)str.length())
 		{
 			if(!isspace((char)str[StartIndex]))
 				break;
@@ -56,7 +56,7 @@ namespace FlamingTorch
 			StartIndex++;
 		};
 
-		if(StartIndex >= str.length())
+		if(StartIndex >= (int32)str.length())
 			return std::string();
 
 		while (EndIndex >= 0)
